Improvements to tasks and questions
There is now a view for showing all questions and their answers

The project view now shows one question's answer.

The project view now shows tasks from the available tasks view. This
view only shows tasks not completed and also not dependant on any tasks
that are not done.

The project view also has a command to split a task into smaller tasks.
This creates a bunch of tasks that all block the task they're breaking
down. When using this to break down a task, you'll have another chance
to assess that you've broken it down properly after completing the
broken down pieces